# session4/api/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"DbUser": {
				"properties": {
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"email": {
						"type": "string"
					},
					"pass": {
						"type": "string"
					},
					"verified": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"user_id",
					"email",
					"pass",
					"verified",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DbTodo": {
				"properties": {
					"todo_id": {
						"type": "number",
						"format": "double"
					},
					"user_id": {
						"type": "number",
						"format": "double"
					},
					"value": {
						"type": "string"
					},
					"completed": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"user": {
						"$ref": "#/components/schemas/DbUser"
					}
				},
				"required": [
					"todo_id",
					"user_id",
					"value",
					"completed",
					"created"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TodoUpdateBody": {
				"properties": {
					"contentDirty": {
						"type": "string"
					},
					"completed": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "session_4_backend_leftovers",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/user/verify/{token}": {
			"get": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "emailDirty",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "shaPass",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/confirmation/{user_id}": {
			"get": {
				"operationId": "Confirmation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"verified": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"verified",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"token",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "emailDirty",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "shaPass",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/todos/add": {
			"post": {
				"operationId": "TodoAdd",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "content",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/todos/list": {
			"post": {
				"operationId": "TodoList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"todos": {
											"items": {
												"$ref": "#/components/schemas/DbTodo"
											},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"todos",
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/todos/remove": {
			"post": {
				"operationId": "TodoRemove",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "todo_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/todos/update": {
			"post": {
				"operationId": "TodoUpdate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"success"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "todo_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TodoUpdateBody"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}